requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
matplotlib>=3.7.0
networkx>=3.1
reportlab>=4.0.0
fpdf2>=2.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
axios>=1.6.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pgmpy>=0.1.23
fedml>=0.8.0
redis>=5.0.0
celery>=5.3.0
pydantic>=2.0.0
